+++ bee-ledger/src/workers/snapshot/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Module containing snapshot operations.
+
+pub(crate) mod download;
+pub(crate) mod error;
+pub(crate) mod import;
+pub(crate) mod worker;
+
+pub mod config;